refactor(api): nest services by logical blocks
