+++ apps/api/src/loaders/authentication.ts
@@ -1,6 +1,6 @@
 import type { Express } from 'express';
 import passport from 'passport';
-import jwtStrategies from '@api/services/auth/strategies';
+import jwtStrategies from '@api/services/core/auth/strategies';
 
 export default (app: Express): void => {
   app.use(passport.initialize());
