+++ apps/api/src/services/admin/images/as-served.service.ts
