+++ apps/api/src/services/admin/images/guide-image.service.ts
