+++ apps/api/src/services/admin/images/image-map.service.ts
