+++ apps/api/src/services/admin/images/processed-image.service.ts
