+++ apps/api/src/services/admin/images/source-image.service.ts
