+++ apps/api/src/services/admin/index.ts
@@ -1,3 +1,7 @@
+export * from './data-export';
+export * from './images';
 export * from './survey.service';
+export * from './user.service';
 
 export { default as adminSurveyService } from './survey.service';
+export { default as userService } from './user.service';
