+++ apps/api/src/services/core/auth/authentication.service.ts
