+++ apps/api/src/services/core/auth/jwt.service.ts
