+++ apps/api/src/services/index.ts
@@ -1,17 +1,6 @@
 export * from './admin';
-export * from './auth';
-export * from './data-export';
+export * from './core';
 export * from './foods';
-export * from './images';
-export * from './pusher';
-export * from './redis';
 export * from './survey.service';
-export * from './user.service';
 
-export { default as Filesystem } from './filesystem';
-export { default as logger } from './logger';
-export { default as Mailer } from './mailer';
-export { default as Pusher } from './pusher';
-export { default as Scheduler } from './scheduler';
 export { default as surveyService } from './survey.service';
-export { default as userService } from './user.service';
